// app/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ROLE = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type ROLE = (typeof ROLE)[keyof typeof ROLE]


export const GENDER = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  UNISEX: 'UNISEX'
} as const

export type GENDER = (typeof GENDER)[keyof typeof GENDER]


export const ORDER_STATUS = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
  SHIPPED: 'SHIPPED',
  COMPLETED: 'COMPLETED',
  REFUNDED: 'REFUNDED'
} as const

export type ORDER_STATUS = (typeof ORDER_STATUS)[keyof typeof ORDER_STATUS]


export const PAYMENT_METHOD = {
  CASH_ON_DELIVERY: 'CASH_ON_DELIVERY',
  BANK_TRANSFER: 'BANK_TRANSFER',
  VIRTUAL_ACCOUNT: 'VIRTUAL_ACCOUNT',
  QRIS: 'QRIS',
  E_WALLET_DANA: 'E_WALLET_DANA',
  E_WALLET_OVO: 'E_WALLET_OVO',
  E_WALLET_GOPAY: 'E_WALLET_GOPAY',
  E_WALLET_SHOPEEPAY: 'E_WALLET_SHOPEEPAY',
  CREDIT_CARD: 'CREDIT_CARD',
  DEBIT_CARD: 'DEBIT_CARD'
} as const

export type PAYMENT_METHOD = (typeof PAYMENT_METHOD)[keyof typeof PAYMENT_METHOD]